refactor: split memos service into separate files with comprehensive test coverage for better maintainability
