+++ internal/memos/usecases/service_operation_list_memos_test.go
@@ -1,0 +1,56 @@
+package usecases
+
+import (
+	"context"
+	"net/http"
+	"testing"
+	"time"
+)
+
+func TestService_ListMemos_Normal(t *testing.T) {
+	client := &mockHTTPClient{
+		doFunc: func(r *http.Request) (*http.Response, error) {
+			if r.Method != http.MethodGet {
+				t.Fatalf("method = %s, want %s", r.Method, http.MethodGet)
+			}
+			if r.URL.Path != "/api/v1/memos" {
+				t.Fatalf("path = %s, want /api/v1/memos", r.URL.Path)
+			}
+			query := r.URL.Query()
+			if got := query.Get("pageSize"); got != "15" {
+				t.Fatalf("pageSize = %s, want 15", got)
+			}
+			if got := query.Get("pageToken"); got != "next-token" {
+				t.Fatalf("pageToken = %s, want next-token", got)
+			}
+			if got := query.Get("state"); got != "NORMAL" {
+				t.Fatalf("state = %s, want NORMAL", got)
+			}
+			if got := query.Get("orderBy"); got != "update_time desc" {
+				t.Fatalf("orderBy = %s, want update_time desc", got)
+			}
+			return jsonResponse(http.StatusOK, `{"memos":[{"name":"memos/1"},{"name":"memos/2"}],"nextPageToken":"next-token","totalSize":2}`), nil
+		},
+	}
+
+	service := NewService(ServiceOptions{
+		BaseURL:    "https://memos.example.com",
+		APIToken:   "token",
+		Timeout:    time.Second,
+		HTTPClient: client,
+	})
+
+	result, err := service.ListMemos(
+		context.Background(),
+		15,
+		"next-token",
+		"NORMAL",
+		"update_time desc",
+	)
+	if err != nil {
+		t.Fatalf("ListMemos() error = %v", err)
+	}
+	if len(result.Memos) != 2 {
+		t.Fatalf("len(memos) = %d, want 2", len(result.Memos))
+	}
+}
